Stop skipping output values in MarshalProperties
I recently introduced the `DontSkipOutputs` option, to maintain the
default (dubious) behavior of skipping outputs, with a new option to
not do that in the places where we don't want to skip outputs.

However, it turns out we'd pretty much need to specify `DontSkipOutputs`
everywhere, so it no longer makes sense to have this as an option.
Skipping outputs to begin with doesn't make much sense as it wasn't also
skipping `Computed`.