Check in default vscode settings
Currently this is just setting go.buildTags to "all" so the language
specific tests are included in build and test commands by default.

Missing this leads to the confusing (at least to new people) behaviour
of clicking run test on a test method and vscode then saying there is no
test to run.