Return output value itself when known, not a secret, and no deps
Plain values set on properties typed as `Input<T>` are eagerly converted to `Output<T>` inside the `Input<T>`. To avoid these from being serialized as output values, when an output is known, not a secret, and has no deps, just return the value itself rather than an output value.