[codegen/go] Simplify object default applies.
The overall goal of this change is to avoid unnecessary conversions
between pointer and non-pointer output types. The use of these
conversions requires that we always generate the pointer variant of an
object type, which impedes #7943's ability to omit these types. Instead,
we can let the return type of the callback match the argument type and
return the dereferenced result of `foo.Defaults()` if necessary.

These changes also simplify the generated code somewhat in order to
avoid introducing additional variables.