[codegen/go] Fix ElementType for nested collection types. (#8535)
The element types for nested collection types are currently incorrect.

For arrays, we generate an element type of `[]FooArgs`, where
`FooArgs` is an args type for `FooInput`. Instead, the element type
should be `[]Foo`, where `Foo` is the element type of `FooInput`.

For maps, we generate an element type of `FooArgs`. Instead, the element
type should be `map[string]Foo`, where `Foo` is the element type of
`FooInput`.

These changes correct the element types.