[sdk/go] Fix copyInputTo for enums. (#8590)
Rather than setting the destination to a value of one of the `pulumi.`
primitive types, set the destination to a value of the appropriate args
type. For example, consider the following definitions:

```
type StringEnumInput interface {
	Input

	ToStringEnumOutput() StringEnumOutput
	ToStringEnumOutputWithContext(context.Context) StringEnumOutput
}

...

type StringEnum string

...

type StringEnumInputArgs struct {
	Value StringEnumInput `pulumi:"value"`
}

func init() {
	RegisterInputType(reflect.TypeOf((*StringEnumInput)(nil)).Elem(), StringEnum(""))
}

```

The original code would have attempted to set
`StringEnumInputArgs.Value` to a value of type `pulumi.String`, but
`pulumi.String` does not implement `StringEnumInput`. The fixed code
creates a value of type `StringEnum`, sets its value to whatever is
present in the object to deserialize, and then sets
`StringEnumInputArgs.Value` to the deserialized value.