Don't crash on null inputs for value types in IOutputCompletionSource.SetValue(OutputData<object?> data)
SetValue correctly annotates that the object held in data might be null.
But then ignores that annotation and tries to cast to `T`, if the object
is null this is fine for reference types which just stay null, but for
value types this throws a NullReferenceException. Simple fix to just use
`default(T)` for when `data.Value` is null, and only cast when non-null.

This has the same behaviour as before for reference types (where
`default(T)` is null) but means that value types take their default
value rather than throwing.

Fixes #8625