[backend/filestate] Allow preview on locked stack
The httpstate backend allows previews to proceed even while an update is in progress.  This is potentially problematic as the preview may be relative to a partial state of the stack, but ensures that previews will not be blocked on stacks that have long running updates (for example, to allow for concurrent PR jobs to preview changes).  This behaviour has been consistent ~forever for the httpstate backend.

In the filestate backend, we recently introduced locking, using a quite different (more coarse-grained) approach. As part of this implementation, preview was added to the list of operations that require an exclusive lock on the stack.

For consistency, we should loosen this so that preview behaves the same relative to state locking in the filestate and httpstate backends.

In the future, we may well want to tighten this up for both backends, with some additional user controls.  Also, when the update plans feature lands shortly, that will provide some additional helpful guarantees that a previous preview was not accidentally relative to a partial state.

Fixes #8609.