[nodejs/sdk] GetRequiredPlugins: Return plugins even when there're errors (#8699)
The implementation of `GetRequiredPlugins` in the Node.js language host will return an empty set of plugins if it encounters an errors reading `package.json` inside a subdirectory. The intended behavior is to return any plugins that were found (even if there were some errors in the process) and to log any errors that occurred along the way. This change fixes the implementation to behave as intended.