[codecov] Fix project names + paths.
Per the [Codecov documentation](), changes to a repository's
configuration may not be reflected if the configuration is invalid.
There is a [documented endpoint](https://docs.codecov.com/docs/codecov-yaml#validate-your-repository-yaml)
that can be used to validate Codecov configuration, which indicated some
problems with our config. These changes correct those issues and add a
Makefile target that can be used to validate the config when making
changes.