Accept `*` as an index in property paths (#8743)
* Accept `*` as an index in property paths

`PropertyPaths` are used as path descriptors with wildcards (`*`). Users
know they can use `root["*"]` to wildcard on any entry in a map. Because
strings and numbers can both be used to index (map vs array), our users
expect to be able to use `root[*]` as a wildcard for an array. We should
parse `[*]` so wildcards work as expected. Internally, this will be
syntactic sugar for `["*"]`, but with much less user confusion.

* Update CHANGELOG_PENDING.md

* Expand simple (`*`) test cases