[python/sdk] Disallow invalid provider option
When a resource R is created with `ResourceOptinos(provider=P)`,
and when P.pkg disagrees with the package providing R, do not
create R using P. Instead create R with the provider it finds in
`providers` (or default) and pass `P` as an overriding member of
`self._providers` to any child resources.

This makes `ResourceOptions(provider=P)` the same as
`ResourceOptions(providers=[P])`. This is desirable since
`ResourceOptinos(providers=[P]).merge(None) ==
ResourceOptions(provider=P)` already.

This is progress towards the reasonable expectation that
`ResourceOptions(**kargs).merge(None) == ResourceOptions(**kargs)`.