[sdk/dotnet] Fix inconsistent automation test
This test was flakey for me when building on Windows & *nix using different versions of the dotnet
framework or depending on whether I had an up to date Pulumi version installed. The regex match is
to blame.

On old versions during CI when we do build a version locally, we expect a message on stdout advising
that a new version is available. If the user hasn't built the language provider locally or has the
current version built, we expect an empty string.

Rather than use regex, which had inconsistent matching behavior and as usual, caused two problems,
we just test if the string is nonempty first.

Once and for all, this test resolves this test passing the matrix consisting of:

 Pulumi version   |    OS   | .NET SDK
 ---------------- | ------- | --------
 3.21.1 (old)     | Windows | 3.1
 3.21.1 (old)     | Windows | 6.0
 3.21.1 (old)     | Linux   | 3.1
 3.21.1 (old)     | Linux   | 6.0
 3.22.2 (current) | Windows | 3.1
 3.22.2 (current) | Windows | 6.0
 3.22.2 (current) | Linux   | 3.1
 3.22.2 (current) | Linux   | 6.0