Add support for overriding plugin download URLs
This commit adds a variable that can be set with a linker flag to override the plugin download URL for matching plugin names.

For example, compiling the Pulumi CLI with the following linker flag:

- `-X [module path]/workspace.pluginDownloadURLOverrides=^foo.*=https://foo,^bar.*=https://bar`

Would make it so that plugin names that start with "foo" are downloaded from https://foo and names that start with "bar" are downloaded from https://bar.