[plugins] Fix plugin reinstallation. (#8892)
Plugin installation has an early return path in the case that the plugin
was already installed. For normal plugin installation, this is okay and
saves some work. For reinstallation, however, this defeats the purpose
by causing installation to return before doing any actual work.

With these changes, we can better accommodate developer scenarios
without requiring that plugins are available on the PATH. For example,
the following script can be used to install a plugin under development:

    tar -c [plugin files] | gzip > plugin.tgz
    pulumi plugin install [kind] [name] [version] -f ./plugin.tgz --reinstall