Allow encrypting the `id` property for a resource
We already handled id correctly when it was an output. We just didn't
handle it correctly for the special ID field in a serialized resource.

When we serialize a resource, we now check if `"id"` is a secret output.
If so, we set `coreapi.ResourceV3.ID` to `resource.SecretSig`. We
reverse this process when deserializing.

This solution is sound as long as
- `stack.SerializeResource` is the only place where resources are
serialized.
- `stack.DeserializeResource` is the only place where resources are
deserialized.
- We can assume that `resource.SecretSig` is not used as an ID.