Fix (Read,Invoke) denied default provider handling
When denying default providers was added, we had no special handling for
Reads and Invokes. This lead to confusing error messages. The fix (#8853)
involved checking on invokes. This check didn't apply to several types
of calls (Read) as well as blocking invokes with providers applied.

This PR fixes the logic to only deny providers when they are default
providers.

It also pushes the change into `getProviderFromSource`, which ensures
that this behavior is handled the same way (and correctly) for both
Invokes and Reads.