Move `StackIdentifier` to httpstate.Backend
The stack method just called into the backend method anyway and this lets us remove a few more type tests where we can just check for httpstate.Backend not for that and httpstate.Stack.

Eventually all StackRefernces will be the same so there won't even be any chance of an error from passing the wrong type of StackReference to StackIdentifier.