Support pointers in pulumi.All
In awaitInputs we didn't correctly handle the case of the input value being a pointer and the resoled destination being `interface{}`. We now handle this by allocating a new pointer of the same type of the input, and assiging that to the resolved location if resolved.Type is `any`.

Else we expect resolved.Type() to be some pointer type that the exisiting logic can handle.