PCL: Use resource aliases when resolving a type (#9238)
* PCL: Use resource aliases when resolving a type

* Add a test

* Fix lint

* Fetch schemas in the makefile

Because we have exposed the `TestProgramCodegen` function to the public,
we can no longer pin specific schema version from that function. In
general, we should avoid mutating global state in tests. This becomes
especially important when the tests are used by other packages.

Using the makefile to setup global state both simplifies how we fetch
schemas and ensures that the fetches are intentional.

* Fix path

* remove unused imports

* Add comments