[codegen/go] Add support for specifying internal dependencies (#9183)
There are situations where an MLC provider may use other resource providers inside its implementation. Such dependencies are not exposed through public APIs and aren't used at all from the generated Go SDK. Since these dependencies aren't used in the Go SDK, there's no way to specify them in the SDK's `go.mod`, as they'd be removed by `go mod tidy`. This is problematic for plugin discovery. Since the dependencies aren't able to be listed in the SDK's `go.mod`, the associated plugins will not be installed automatically for users of the SDK.

This change introduces a new Go schema option (`InternalDependencies`) that can be used to specify a set of packages that will be emitted as blank imports in the generated SDK so that the associated dependencies in `go.mod` will not be removed by `go mod tidy`.