Revert "Strip Aliases from state (#9275)" (#9282)
This reverts commit 17068e9b4907cc22acc65c3f80aca2f6aa8caa9f.

Turns out NormalizeURNReferences needs this in the state to fix up URNs while the deployment is running. It feels like we should be able to either thread this information through to the snapshot manager another way but it's not obvious how. It's also tricky to test because snapshot code differs massively in unit tests compared to proper runs.