[engine] Clear pending operations with refresh. (#8435)
* [engine] Clear pending operations with refresh.

Just what it says on the tin. This is implemented by moving the check
for pending operations in the last statefile into the deployment
executor and making it conditional on whether or not a refresh is being
performed (either via `pulumi refresh` or `pulumi up -r`). Because
pending operations are not carried over from the base statefile, this
has the effect of clearing pending operations if a refresh is performed.

Fixes #4265.

* CL

* nil ref

* adjust a test

* Return nil for the plan when executing deployment

* fix lifecycle test

* parallel test for TestRefreshWithPendingOperations

* preserve pending create operations

Co-authored-by: Zaid Ajaj <zaid.naom@gmail.com>