deps: Upgrade to Cobra v1.4.0
This commit upgrades the usage of Cobra to [v1.4.0][1], which reduces
the dependency set by removing the dependency on Viper. The upgrade
happens across both `sdk` and `pkg`.

Full details on the upgrades upstream in Cobra are available via the
release notes, but [2] contains further motivating examples of reduced
dependencies via separation of the (rarely used) Cobra bootstrapping CLI
and the library itself.

[1]: https://github.com/spf13/cobra/releases/tag/v1.4.0

[2]: https://github.com/spf13/cobra/issues/1138