[cli] Render changes to text properties as diffs.
Instead of simply rendering the old and new values for textual
properties, compute the diff between the old and new values and render
that. For single-line values, the diff is computed and rendered
character-wise, with simplification applied to avoid "small" diffs. For
multi-line values, the diff is computed and rendered linewise (as we
already to for textual assets).

Fixes #9136.