Add NewUniqueName
NewUniqueHexV2 tried to generate deterministic names via sequence
numbers, but this was insufficent to generate names that worked for many
scenarios. We're going to try to generate names based on more sources of
data, but there's no need to keep updating providers and provider code
each time we change the data set. Instead we're going to pass the byte
slice to Check and use that as a random source, the engine will then be
free to change how those random bytes get generated (including possibly
just generating them completly randomly)

This adds a new function for generating unique names that will work with
that idea. It takes a byte slice and uses that to generate a random
number sequence to fill in the random part of the name. This new method
also takes a "charset" option as currently there's a number of resources
spread across different providers that have their own name generation
code due to them needing a different charset to [a-f0-9].

At a later point we can probably remove NewUniqueHexV2. While it's
technically part of our exported public API nothing should be using it
(and a search across github doesn't show any hits).