Fix race condition in reading EngineEvents from tailed logs
We we're seeing the odd test fail because preview returned that len(summaryEvents) was 0. This was because everything about this code is async and didn't do any synchronisation so it was pure luck that the file was read, the message was sent to the channel, and then the channel goroutine read the message and appened it to summaryEvents in time for the len check to pass.

Now we close off the file watcher and wait for the channels and goroutines to flush and signal done before checking the length of summaryEvents.