Stop using Get/SetRootResource in nodejs
Just use globalThis to make the stack resource a global rather than a module variable. This ensures that even if we end up with multiple Pulumi modules they can all grab the stack object and read URN off it.

This is also a temporary fix, we're going to move a lot of the parenting logic to the engine so SDKs won't need this global state.