Further explain kwargs for Python stack funcs
The *_stack functions in the pulumi.automation package take keyword
arguments that control where the program used in the stack comes
from. In particular, the keyword argument work_dir (and absence of
program and project_name) indicates that the program should be
constructed using the given directory. There's _also_ a field with the
same name accepted in the opts argument, leading to confusion in at
least one case (see https://github.com/pulumi/pulumi/issues/8436).

This commit tries to mitigate the confusion by making the docstrings for
those functions more explicit about which arguments are required and
ignored in each case.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>