Allow a path to a file when cons'ing an archive
It's legitimate to construct an archive with a single file (rather than
a directory) by mentioning the file's name:

    zip handler.zip handler/main

In fact, that's what the AWS Lambda docs suggest:
https://docs.aws.amazon.com/lambda/latest/dg/golang-package.html

This commit lets you do the equivalent when making an
`asset.FileArchive`. It needs a little rearrangement of
directoryArchiveReader so that the individual paths are relative to a
base rather than sharing a prefix; otherwise, a single-file archive (in
which the name is the entire file path) cannot be constructed. I've
sign-posted this change by renaming the helper struct in question to
`filesystemArchiveReader`.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>