Include invoke.ts in tsconfig.files
.. and consequently, the type therein `InvokeOptions`, in generated
docs. All functions in Pulumi SDKs take an InvokeOptions-typed argument,
and all the generated documentation for those functions has a broken
link because it's not part of the NodeJS SDK documentation generated for
pulumi/pulumi.

The version of TypeDoc used is quite old, and I suspect it consults
`tsconfig.json` to find the source files (evidence: someone complaining
that it does not do this well enough:
https://github.com/TypeStrong/typedoc/issues/1515). I have
conservatively just added the file in question to `files`, since:

 - I'm not sure if .files should list _all_ the files
 - Upgrading TypeDoc would surely have many unintended consequences

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>